+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add support for Python 3.12
 * Add support for Django 5.0.
 * Introduce support for floating point size QR code modules.
+* Introduce embedded Base64 SVG image.
 
 ## 3.1.2 (2023-04-10)
 * Fix bug in `WifiConfig` data class (#43).